+++ backend/lcfs/db/models/Transfer.py
@@ -13,8 +13,10 @@ class Transfer(BaseModel, Auditable, EffectiveDates):
     from_organization_id = Column(Integer, ForeignKey('organization.organization_id'))
     to_organization_id = Column(Integer, ForeignKey('organization.organization_id'))
     transaction_id = Column(Integer, ForeignKey('transaction.transaction_id'))
+    agreement_date = Column(DateTime, comment="Agreement date of the transfer")
     transaction_effective_date = Column(DateTime, comment="transaction effective date")
     price_per_unit = Column(Integer, comment="Price per unit")
+    quantity = Column(Integer, comment="Quantity of units")
     # compliance_period = Column(Integer, )
     comment_id = Column(Integer, ForeignKey('comment.comment_id'))
     transfer_status_id = Column(Integer, ForeignKey('transfer_status.transfer_status_id'))
