+++ frontend/src/views/Transfers/AddTransfer/TransferForm.test.jsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach } from 'vitest'
 import { render, screen, cleanup, fireEvent } from '@testing-library/react'
 import { useForm, FormProvider } from 'react-hook-form'
-import TransferInputs from './TransferDetails'
+import TransferInputs from './TransferForm'
 import { BrowserRouter as Router } from 'react-router-dom'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ThemeProvider } from '@mui/material/styles'
